refactor: Add PrintUtils to separate print methods
Added PrintUtils to separate print methods from Game class.
