+++ src/main/java/chess/Game.java
@@ -1,11 +1,8 @@
 package chess;
 
-import pieces.Color;
-import pieces.Piece;
-
-import java.util.Collections;
-import java.util.List;
 import java.util.Scanner;
+
+import static util.PrintUtils.showScore;
 
 public class Game {
     private static final String START = "start";
@@ -40,28 +37,6 @@ public class Game {
             } else {
                 System.out.println(command + " not supported.");
             }
-        }
-    }
-
-    static void showScore(Board board) {
-        double whiteScore = board.calculatePoint(Color.WHITE);
-        double blackScore = board.calculatePoint(Color.BLACK);
-        System.out.println("✓ White score : " + whiteScore);
-        System.out.println("✓ Black score : " + blackScore);
-        showScoreByPiece(board);
-    }
-    static void showScoreByPiece(Board board){
-        List<Piece> whitePieces = board.findPiecesByColor(Color.WHITE);
-        List<Piece> blackPieces = board.findPiecesByColor(Color.BLACK);
-        Collections.sort(whitePieces);
-        Collections.sort(blackPieces);
-        System.out.println("\n[White Player's result]");
-        for (Piece piece : whitePieces) {
-            System.out.println(piece.getType() + " : " + piece.getType().getDefaultPoint());
-        }
-        System.out.println("\n[Black Player's result]");
-        for (Piece piece : blackPieces) {
-            System.out.println(piece.getType() + " : " + piece.getType().getDefaultPoint());
         }
     }
